+++ odoo/songs/install/data_full.py
@@ -198,6 +198,21 @@ def import_pricelist(ctx):
 
 
 @anthem.log
+def import_invoices_supplier(ctx):
+    """ Importing invoices supplier from csv """
+    model = ctx.env['account.invoice'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/invoice_supp_head.csv',
+             model)
+    model_item = ctx.env['account.invoice.line'].with_context({
+        'tracking_disable': True,
+    })
+    load_csv(ctx, 's3://prod-sf-odoo-data/install/invoice_supp_line.csv',
+             model_item)
+
+
+@anthem.log
 def main(ctx):
     """ Loading full data """
     import_users(ctx)
@@ -221,4 +236,5 @@ def main(ctx):
     import_bank_account(ctx)
     import_sales_order(ctx)
     import_bom(ctx)
+    import_invoices_supplier(ctx)
     return